+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/dto/CabinetDTO.java
@@ -2,12 +2,13 @@ package com.doubleslash.fifth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
 public class CabinetDTO {
 	
-	private int aid;
+	private Long aid;
 	
 	private String image;
