+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/repository/ReviewLoveRepository.java
@@ -14,15 +14,15 @@ public interface ReviewLoveRepository extends JpaRepository<ReviewLove, Integer>
 
 	// 찜한 리뷰 확인
 	@Query(value = "select id, rid from ReviewLove where id = ?1 and rid = ?2", nativeQuery = true)
-	public ReviewLove findByIdRid(int id, int rid);
+	public ReviewLove findByIdRid(Long id, Long rid);
 	
 	@Modifying
 	@Transactional
 	@Query(value = "insert ignore into ReviewLove(id, rid) values(?1, ?2)", nativeQuery = true)
-	public int insert(int id, int rid);
+	public int insert(Long id, Long rid);
 	
 	@Modifying
 	@Transactional
 	@Query(value = "delete from ReviewLove where id=?1 and rid=?2", nativeQuery = true)
-	public int delete(int id, int rid);
+	public int delete(Long id, Long rid);
 }
