+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/service/GYJ_RecommendService.java
@@ -45,7 +45,7 @@ public class GYJ_RecommendService {
 	private final int CB_WEIGHT = 10;
 	
 	//추천 로직
-	public void createRecommend(RecommendDTO rec, int id){
+	public void createRecommend(RecommendDTO rec, Long id){
 		double minAbv = rec.getMinAbv();
 		double maxAbv = calcAbv(rec.getMaxAbv());
 		int abvWeight = calcAbvWeight(rec.getMaxAbv() - rec.getMinAbv());
@@ -62,7 +62,7 @@ public class GYJ_RecommendService {
 		//int celeb; 인지도
 		
 		//스코어링 Map
-		HashMap<Integer, RecommendDTO.map> score = new HashMap<>();
+		HashMap<Long, RecommendDTO.map> score = new HashMap<>();
 		
 		RecommendDTO.liquor liquor = rec.getAlcohol().getLiquor();
 		if(liquor != null) {
@@ -74,8 +74,8 @@ public class GYJ_RecommendService {
 			
 			for(int i = 0; i < liquorCompTarget.size(); i++) {
 				target = liquorCompTarget.get(i); //타겟 설정
-				int aid = target.getAid();
-				score.put(aid, new RecommendDTO.map("양주", 0)); //스코어링 데이터 삽입
+				Long aid = target.getAid();
+				score.put(aid, new RecommendDTO.map("양주", 0L)); //스코어링 데이터 삽입
 				
 				//맛 스코어링
 				String kind = target.getKind();
@@ -126,8 +126,8 @@ public class GYJ_RecommendService {
 			
 			for(int i = 0; i < wineCompTarget.size(); i++) {
 				target = wineCompTarget.get(i); //타겟 설정
-				int aid = target.getAid();
-				score.put(aid, new RecommendDTO.map("와인", 0)); //스코어링 데이터 삽입
+				Long aid = target.getAid();
+				score.put(aid, new RecommendDTO.map("와인", 0L)); //스코어링 데이터 삽입
 				
 				//맛 스코어링
 				String kind = target.getKind();
@@ -174,8 +174,8 @@ public class GYJ_RecommendService {
 			
 			for(int i = 0; i < beerCompTarget.size(); i++) {
 				target = beerCompTarget.get(i); //타겟 설정
-				int aid = target.getAid();
-				score.put(aid, new RecommendDTO.map("세계맥주", 0)); //스코어링 데이터 삽입
+				Long aid = target.getAid();
+				score.put(aid, new RecommendDTO.map("세계맥주", 0L)); //스코어링 데이터 삽입
 				
 				//맛 스코어링
 				String kind = target.getKind();
@@ -220,20 +220,20 @@ public class GYJ_RecommendService {
 
 		}
 		
-		List<Entry<Integer, RecommendDTO.map>> entries = new ArrayList<Entry<Integer, RecommendDTO.map>>(score.entrySet());
+		List<Entry<Long, RecommendDTO.map>> entries = new ArrayList<Entry<Long, RecommendDTO.map>>(score.entrySet());
 		
 		for(int i = 0; i < 10; i++) {
-			Collections.sort(entries, new Comparator<Entry<Integer, RecommendDTO.map>>() {
-				public int compare(Entry<Integer, RecommendDTO.map> obj1, Entry<Integer, RecommendDTO.map> obj2)
+			Collections.sort(entries, new Comparator<Entry<Long, RecommendDTO.map>>() {
+				public int compare(Entry<Long, RecommendDTO.map> obj1, Entry<Long, RecommendDTO.map> obj2)
 				{
-					return Integer.compare(obj2.getValue().getRecScore(), obj1.getValue().getRecScore());
+					return Long.compare(obj2.getValue().getRecScore(), obj1.getValue().getRecScore());
 				}
 			});	
 			
-			Entry<Integer, RecommendDTO.map> entry = entries.get(i);
-			int aid = entry.getKey();
+			Entry<Long, RecommendDTO.map> entry = entries.get(i);
+			Long aid = entry.getKey();
 			String category = entry.getValue().getCategory();
-			int recScore = entry.getValue().getRecScore();
+			Long recScore = entry.getValue().getRecScore();
 			
 			for(int j = i+1; j < entries.size(); j++) {
 				RecommendDTO.map temp = entries.get(j).getValue();
