+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/storage/BeerStorage.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class BeerStorage {
 
-	int aid;
+	Long aid;
 	
 	int lowestPrice;
